orders/requirements.txt
amqp==5.2.0
asgiref==3.7.2
attrs==23.2.0
billiard==4.2.0
celery==5.3.6
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.4.0
cryptography==41.0.7
defusedxml==0.8.0rc2
Django==4.2.6
django-baton==2.9.1
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
drf-spectacular==0.27.0
easy-thumbnails==2.8.5
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
kombu==5.3.4
oauthlib==3.2.2
packaging==23.2
pillow==10.2.0
platformdirs==3.10.0
pluggy==1.3.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
pytest==7.4.4
pytest-cov==4.1.0
pytest-django==4.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.32.1
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.17.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
tzdata==2023.3
ujson==5.8.0
uritemplate==4.1.1
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.12
